backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pyyaml==6.0.1
httpx==0.27.0
python-multipart==0.0.9
jsonref>=1.0.0
openapi-spec-validator>=0.5.0
reportlab>=4.0.0
openai>=1.0.0
